feat: Integrate ESP32P4 hardware H.264 encoder and HTTP-based WiFi provisioning
Major enhancements for ESP32P4 network camera system:

## H.264 Hardware Encoding
- Add ESP32P4 hardware H.264 encoder integration with esp_h264 component
- Implement ESP_H264_RAW_FMT_O_UYY_E_VYY format conversion for hardware encoding
- Add software simulation fallback when hardware encoder is unavailable
- Fix memory management issues with safe buffer allocation and deallocation
- Optimize memory usage with reduced buffer sizes for initial testing

## WiFi Provisioning
- Implement complete HTTP-based WiFi provisioning server
- Add responsive web interface for WiFi configuration (http://192.168.4.1)
- Support AP mode provisioning with automatic STA mode switching
- Integrate WiFi credential storage and connection management
- Add proper state handling for provisioning workflow

## Memory Management
- Fix heap corruption issues in H.264 encoder buffer management
- Add null pointer initialization and safe memory cleanup
- Implement proper error handling and fallback mechanisms
- Reduce memory footprint with optimized buffer configurations

## Component Integration
- Add ESP32P4-specific components: esp_h264, esp_wifi_remote, esp_hosted
- Update component dependencies and build configuration
- Ensure compatibility with ESP32P4 hosted WiFi architecture

The system now provides a complete solution for ESP32P4 network cameras with
hardware-accelerated video encoding and user-friendly WiFi setup.

+++ main/task_manager.cpp
@@ -188,12 +188,12 @@ void TaskManager::printTaskStatus() const {
     auto tasks = getAllTasks();
     for (const auto& task_pair : tasks) {
         const TaskInfo& info = task_pair.second;
-        ESP_LOGI(TAG, "Task: %s | State: %d | Priority: %d | Stack: %d | Watermark: %d",
+        ESP_LOGI(TAG, "Task: %s | State: %d | Priority: %lu | Stack: %lu | Watermark: %lu",
                  info.name.c_str(),
                  (int)info.state,
-                 info.priority,
-                 info.stack_size,
-                 getStackWatermark(info.name));
+                 (unsigned long)info.priority,
+                 (unsigned long)info.stack_size,
+                 (unsigned long)getStackWatermark(info.name));
     }
     
     ESP_LOGI(TAG, "=========================");
@@ -201,8 +201,8 @@ void TaskManager::printTaskStatus() const {
 
 void TaskManager::printMemoryUsage() const {
     ESP_LOGI(TAG, "=== Memory Usage Report ===");
-    ESP_LOGI(TAG, "Free heap: %d bytes", esp_get_free_heap_size());
-    ESP_LOGI(TAG, "Minimum free heap: %d bytes", esp_get_minimum_free_heap_size());
+    ESP_LOGI(TAG, "Free heap: %lu bytes", (unsigned long)esp_get_free_heap_size());
+    ESP_LOGI(TAG, "Minimum free heap: %lu bytes", (unsigned long)esp_get_minimum_free_heap_size());
     ESP_LOGI(TAG, "===========================");
 }
 
@@ -223,14 +223,14 @@ void TaskManager::checkStackWatermarks() const {
             uint32_t watermark = uxTaskGetStackHighWaterMark(info.handle);
             uint32_t usage_percent = ((info.stack_size - watermark) * 100) / info.stack_size;
             
-            ESP_LOGI(TAG, "Task '%s': %d/%d bytes used (%d%%)",
+            ESP_LOGI(TAG, "Task '%s': %lu/%lu bytes used (%lu%%)",
                      info.name.c_str(),
-                     info.stack_size - watermark,
-                     info.stack_size,
-                     usage_percent);
+                     (unsigned long)(info.stack_size - watermark),
+                     (unsigned long)info.stack_size,
+                     (unsigned long)usage_percent);
             
             if (usage_percent > 80) {
-                ESP_LOGW(TAG, "Task '%s' stack usage is high: %d%%", info.name.c_str(), usage_percent);
+                ESP_LOGW(TAG, "Task '%s' stack usage is high: %lu%%", info.name.c_str(), (unsigned long)usage_percent);
             }
         }
     }
